+++ Sources/Perception/PerceptionTracking.swift
@@ -218,7 +218,7 @@ public func withPerceptionTracking<T>(
   onChange: @autoclosure () -> @Sendable () -> Void
 ) -> T {
   #if canImport(Observation)
-    if isObservationAllowed, #available(iOS 17, macOS 14, tvOS 17, watchOS 10, *) {
+    if #available(iOS 17, macOS 14, tvOS 17, watchOS 10, *), !isObservationBeta {
       return withObservationTracking(apply, onChange: onChange())
     }
   #endif
